fix: resource kinds selectors
